Reorganizing api helpers
