+++ app/items/[uuid]/page.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Metadata } from "next";
 import PageLayout from "../../src/components/pageLayout/pageLayout";
 import Item from "../../src/components/items/item";
-import { getItemData } from "../../src/utils/api";
+import { getItemData } from "../../src/utils/apiHelpers";
 import { createAdobeAnalyticsPageName } from "../../src/utils/utils";
 import { ItemModel } from "../../src/models/item";
