+++ app/src/utils/apiResponse.test.tsx
@@ -1,80 +1,4 @@
-import { waitFor } from "@testing-library/react";
-import { apiResponse, getItemsCountFromUUIDs } from "./api";
-
-describe("getHomePageData", () => {});
-
-describe("getFeaturedItemData", () => {});
-
-describe("getFeaturedImage", () => {});
-
-describe("getItemData", () => {});
-
-describe("getNumDigitizedItems", () => {});
-
-describe("getItemsCountFromUUIDs", () => {
-  it("should throw an error with a bad request", async () => {
-    (global as any).fetch = jest.fn(() =>
-      Promise.resolve({
-        status: 403,
-        json: () => Promise.resolve({ nyplAPI: { response: "success" } }),
-      })
-    ) as jest.Mock;
-
-    const data = ["uuid1", "uuid2"];
-    await expect(getItemsCountFromUUIDs(data)).rejects.toEqual(
-      new Error("apiResponse: 403 No message")
-    );
-  });
-
-  it("should return an empty object with a successful request but bad data structure", async () => {
-    (global as any).fetch = jest.fn(() =>
-      Promise.resolve({
-        status: 200,
-        json: () => Promise.resolve({ nyplAPI: { response: { counts: {} } } }),
-      })
-    ) as jest.Mock;
-
-    const data = ["uuid1", "uuid2"];
-
-    await waitFor(async () => {
-      expect(await getItemsCountFromUUIDs(data)).toEqual({});
-    });
-  });
-
-  it("should return a map of all the uuids passed with their item count", async () => {
-    (global as any).fetch = jest.fn(() =>
-      Promise.resolve({
-        status: 200,
-        json: () =>
-          Promise.resolve({
-            nyplAPI: {
-              response: {
-                counts: {
-                  count: [
-                    { uuid: { $: "uuid1" }, count_value: { $: "10" } },
-                    { uuid: { $: "uuid2" }, count_value: { $: "45" } },
-                  ],
-                },
-              },
-            },
-          }),
-      })
-    ) as jest.Mock;
-
-    const data = ["uuid1", "uuid2"];
-
-    await waitFor(async () => {
-      expect(await getItemsCountFromUUIDs(data)).toEqual({
-        uuid1: "10",
-        uuid2: "45",
-      });
-    });
-  });
-});
-
-describe("getItemByIdentifier", () => {});
-
-describe("getDivisionData", () => {});
+import { apiResponse } from "./apiResponse";
 
 describe("apiResponse", () => {
   const mockApiUrl = "mockurl.org";
